Cleaned up the lexer.
Namely, I fixed the handling of whitespace and comments. In this language, most whitespace is insignificant, but newlines *are* significant, so they can't be arbitrarily skipped. I've now changed the lexer so that '\n's aren't skipped when skipping comments and whitespace unless we are currently processing a '\n'; in other words, multiple '\n's can be collapsed into one. This also fixes some issues with 'separators' for literals and identifiers. For example '#FFFF0' should not count as a valid hexidecimal token, because after '#FFFF' there is no space/comma/paren/etc.